Refactor video processing signals and tasks for HLS conversion and thumbnail extraction

+++ video_app/signals.py
@@ -1,29 +1,23 @@
-from .models import Video
-from django.dispatch import receiver
-from django.db.models.signals import post_save, post_delete
 import os
-from .tasks import convert_480p
+
 import django_rq
+from django.db.models.signals import post_delete, post_save
+from django.dispatch import receiver
+
+from .models import Video
+from .tasks import process_video
+
 
 @receiver(post_save, sender=Video)
-def video_post_save(sender, instance, created, **kwargs):
-    print(f'Video "{instance.title}" has been saved.')
-
-    if created:
-        print('New object created')
-        queue = django_rq.get_queue('default', autocommit=True)
-        queue.enqueue(convert_480p, instance.video_file.path)
+def enqueue_video_processing(sender, instance, created, **kwargs):
+    """Schedule HLS conversion and thumbnail generation on initial upload."""
+    if not created:
+        return
+    django_rq.get_queue('default').enqueue(process_video, instance.pk)
 
 
 @receiver(post_delete, sender=Video)
 def auto_delete_file_on_delete(sender, instance, **kwargs):
-    """
-    Deletes file from filesystem
-    when corresponding `Video` object is deleted.
-    """
-    if instance.video_file:
-        if os.path.isfile(instance.video_file.path):
-            os.remove(instance.video_file.path)
-
-
-
+    """Remove the uploaded source file when a Video row is deleted."""
+    if instance.video_file and os.path.isfile(instance.video_file.path):
+        os.remove(instance.video_file.path)
